% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HospitalizationInSentHomePatientsValidation.R
\docType{package}
\name{HospitalizationInSentHomePatientsValidation}
\alias{HospitalizationInSentHomePatientsValidation}
\title{HospitalizationInSentHomePatientsValidation}
\description{
HospitalizationInSentHomePatientsValidation
}
